More title adjustments
Trimmed all the logos, so we can size and position them better.

Removed all uses of textpos and positioned everything using tikz instead. (It's not nice to mix textpos and tikz and this allows us to position the logos relative to the grey box.)

Adjusted sizing and spacing of titles. In particular, got rid of the overfull hbox warnings that were plaguing us before.