Fixed compilation of notes
When compiling Beamer notes, the background canvas doesn't get rendered.
However, our background canvas defines the handy "grey background" node,
so that things can be positioned relative to the main part of the slide.
If you do this, the notes don't compile, since the node hasn't been
defined.

Now we redefine the insertslideinnotes command to add the shape,
correctly positioned in the mini version of the slide shown on the
notes.