Fixed longstanding problem with grey background
Grey box was drawn on background canvas and various commands use it for
positioning things on slides (and it's typically a useful way to
position things in your frame content). However, the background canvas
is rendered *after* the frame content. This is usually not a problem,
since the position of the grey background node is available from the
previous slide, but you couldn't use it on the first frame of a
presentation!

Now an invisible version of the grey box is drawn on every slide in the
headline template. It's then available as a node to refer to for
positioning. The coloured version is then drawn (or not drawn, depending
on the frame type) by the background canvas.